quick refactor to move the load function and cleanup
