+++ aggregate/PUMS/pums_0812_1519_demographics.py
@@ -7,13 +7,10 @@ breakdowns]).
 
 from webbrowser import get
 import pandas as pd
-#from aggregate.PUMS.pums_2000_demographics import rename_columns
-from aggregate.aggregation_helpers import demographic_indicators_denom
-from utils.PUMA_helpers import clean_PUMAs, dcp_pop_races
+from aggregate.aggregation_helpers import demographic_indicators_denom, order_multiyr_aggregated_columns, get_category, get_geography_pop_data
+from utils.PUMA_helpers import dcp_pop_races
 from internal_review.set_internal_review_file import set_internal_review_files
-from aggregate.aggregation_helpers import order_multiyr_aggregated_columns, get_category, get_geography_pop_data
-from aggregate.clean_aggregated import rename_columns_demo
-
+from aggregate.load_aggregated import load_clean_pop_demographics
 
 endyear_mapper = {
     "12" : "0812",
@@ -30,21 +27,11 @@ def demographics(geography: str, write_to_internal_review=False) -> pd.DataFrame
         "race": dcp_pop_races,
     }
     clean_data_0812 = load_clean_pop_demographics("12", endyear_mapper["12"])
-
     clean_data_1519 = load_clean_pop_demographics("19", endyear_mapper["19"])
 
-    final_0812 = get_geography_pop_data(clean_data_0812, geography)
-    #print("0812", final_0812)
-    
+    final_0812 = get_geography_pop_data(clean_data_0812, geography)    
     final_1519 = get_geography_pop_data(clean_data_1519, geography)
-    #print("1519", final_1519)
-
     final = pd.concat([final_0812, final_1519], axis=1)
-
-    final.set_index(geography, inplace=True)
-
-    #print(final)
-    print(final.columns)
 
     final = order_multiyr_aggregated_columns(
         df=final,
@@ -62,33 +49,3 @@ def demographics(geography: str, write_to_internal_review=False) -> pd.DataFrame
             "demographics",
         )
     return final 
-
-
-def load_clean_pop_demographics(end_year: int, year: str) -> pd.DataFrame:
-    """Function to merge the two files for the QOL outputs and do some standard renaming. Because
-    these are QOL indicators they remain in the same csv output with columns indicating year"""
-
-    read_excel_arg = {
-        "0812": {
-            "io": "./resources/ACS_PUMS/EDDT_Dem_ACS2008-2012.xlsx",
-            "sheet_name": "Dem08-12",
-            "usecols": "A:HR",
-            "dtype": {"Geog": str},
-        },
-        "1519": {
-            "io": "./resources/ACS_PUMS/EDDT_Dem_ACS2015-2019.xlsx",
-            "sheet_name": "Dem15-19",
-            "usecols": "A:HR",
-            "dtype": {"Geog": str},
-        },
-    }
-
-    df = pd.read_excel(**read_excel_arg[year])
-
-    df.loc[df["Geog"] == "NYC", "Geog"] = "citywide"
-
-    clean_data = rename_columns_demo(df, end_year, year)
-
-    clean_data.rename(columns={"geog": "Geog"}, inplace=True)
-
-    return clean_data
