+++ aggregate/load_aggregated.py
@@ -10,6 +10,8 @@ from utils.PUMA_helpers import get_all_NYC_PUMAs, get_all_boroughs
 from utils.setup_directory import setup_directory
 from os import path
 import pandas as pd
+from aggregate.clean_aggregated import rename_columns_demo
+
 
 # from aggregate.PUMS.count_PUMS_demographics import PUMSCountDemographics
 # from aggregate.PUMS.median_PUMS_demographics import PUMSMedianDemographics
@@ -108,3 +110,32 @@ def load_clean_housing_security_pop_data(name_mapper: dict) -> pd.DataFrame:
     df.loc[df["Geog"] == "NYC", "Geog"] = "citywide"
 
     return df
+
+def load_clean_pop_demographics(end_year: int, year: str) -> pd.DataFrame:
+    """Function to merge the two files for the QOL outputs and do some standard renaming. Because
+    these are QOL indicators they remain in the same csv output with columns indicating year"""
+
+    read_excel_arg = {
+        "0812": {
+            "io": "./resources/ACS_PUMS/EDDT_Dem_ACS2008-2012.xlsx",
+            "sheet_name": "Dem08-12",
+            "usecols": "A:HR",
+            "dtype": {"Geog": str},
+        },
+        "1519": {
+            "io": "./resources/ACS_PUMS/EDDT_Dem_ACS2015-2019.xlsx",
+            "sheet_name": "Dem15-19",
+            "usecols": "A:HR",
+            "dtype": {"Geog": str},
+        },
+    }
+
+    df = pd.read_excel(**read_excel_arg[year])
+
+    df.loc[df["Geog"] == "NYC", "Geog"] = "citywide"
+
+    clean_data = rename_columns_demo(df, end_year, year)
+
+    clean_data.rename(columns={"geog": "Geog"}, inplace=True)
+
+    return clean_data
